+++ aimiaobi-20230801/core/Models/SaveOrUpdateOssConfigRequest.cs
@@ -1,0 +1,40 @@
+// This file is auto-generated, don't edit it. Thanks.
+
+using System;
+using System.Collections.Generic;
+using System.IO;
+
+using Tea;
+
+namespace AlibabaCloud.SDK.AiMiaoBi20230801.Models
+{
+    public class SaveOrUpdateOssConfigRequest : TeaModel {
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>xxx</para>
+        /// </summary>
+        [NameInMap("BucketName")]
+        [Validation(Required=false)]
+        public string BucketName { get; set; }
+
+        /// <summary>
+        /// <b>Example:</b>
+        /// <para>oss-cn-shanghai.aliyuncs.com</para>
+        /// </summary>
+        [NameInMap("EndPoint")]
+        [Validation(Required=false)]
+        public string EndPoint { get; set; }
+
+        /// <summary>
+        /// <para>This parameter is required.</para>
+        /// 
+        /// <b>Example:</b>
+        /// <para>llm-xxx</para>
+        /// </summary>
+        [NameInMap("WorkspaceId")]
+        [Validation(Required=false)]
+        public string WorkspaceId { get; set; }
+
+    }
+
+}